requirements.txt
pycaret.regression
pandas 
streamlit 